Removed unused classes and configured db throughput
