+++ data-azure-cosmos/src/test/java/io/micronaut/data/azure/entities/CosmosPage.java
@@ -19,10 +19,8 @@ import io.micronaut.data.annotation.GeneratedValue;
 import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.MappedEntity;
 import io.micronaut.data.annotation.Relation;
-import io.micronaut.data.cosmos.annotation.Container;
 
 @MappedEntity
-@Container(name = "pages", throughputRequestUnits = 400)
 public class CosmosPage {
     @GeneratedValue
     @Id
